+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGEZRCYMJT7GXDPG83CKW44V"
+        "01KGGYW6AMVPQD0KDH39BHNCGY"
     )
     r"""The unique identifier of the data source"""