+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGEZRCYHY3RT3E5EGGE20QYJ"
+        "01KGGYW6AF51XXSCHQCSEH8B5F"
     )
     r"""The unique identifier of the data source"""