+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGEZRCYKWRQTJ4V6H49MRAND"
+        "01KGGYW6AJNS6CKXSK8VFC7EZ6"
     )
     r"""The unique identifier of the data source"""